fix(misra): 10.4 批1 接力提交 — 中断会话残留机械修复（63 文件，U 后缀/括号）+ 修复 0x8001UU 重复后缀

+++ src/asw/communication_manager/src/Swc_CommunicationManager.c
@@ -103,7 +103,7 @@ STATIC void Swc_CommunicationManager_ProcessRxSignals(void)
     Swc_PduInfoType pdu;
 
     /* Read PDU from RTE */
-    if ((uint8_t)(Rte_Read_PduData(&pdu)) == RTE_E_OK) {
+    if ((uint8_t)((uint8_t)(Rte_Read_PduData(&pdu))) == RTE_E_OK) {
         /* Find or create PDU entry */
         sint16 pduIndex = Swc_CommunicationManager_FindRxPdu(pdu.pduId);
 
@@ -126,7 +126,7 @@ STATIC void Swc_CommunicationManager_ProcessRxSignals(void)
             /* Simplified - in real implementation, use signal database */
             for (i = 0; i < swcCommManager.numSignals; i++) {
                 if ((swcCommManager.signals[i].signalId >= (pdu.pduId * 10U)) &&
-                    (swcCommManager.signals[i].signalId < ((pdu.pduId + 1U) * 10))) {
+                    (swcCommManager.signals[i].signalId < ((pdu.pduId + 1U) * 10U))) {
                     /* Update signal value from PDU data */
                     swcCommManager.signals[i].value = 0;
                     for (j = 0; (j < 8U) && (j < pdu.length); j++) {
@@ -164,7 +164,7 @@ STATIC void Swc_CommunicationManager_ProcessTxSignals(void)
         pdu->timestamp = Rte_GetTime();
 
         /* Send PDU via RTE */
-        if ((uint8_t)(Rte_Write_PduData(pdu)) == RTE_E_OK) {
+        if ((uint8_t)((uint8_t)(Rte_Write_PduData(pdu))) == RTE_E_OK) {
             swcCommManager.statistics.txPdus++;
         } else {
             swcCommManager.statistics.txErrors++;
@@ -487,7 +487,7 @@ Rte_StatusType Swc_CommunicationManager_SendPdu(const Swc_PduInfoType* pdu)
     swcCommManager.txPdus[pduIndex] = *pdu;
 
     /* Send immediately via RTE */
-    if ((uint8_t)(Rte_Write_PduData(pdu)) == RTE_E_OK) {
+    if ((uint8_t)((uint8_t)(Rte_Write_PduData(pdu))) == RTE_E_OK) {
         swcCommManager.statistics.txPdus++;
         return RTE_E_OK;
     } else {
